+++ test/src/gen_KAT.c
@@ -5,9 +5,9 @@
 #include <stddef.h>
 #include <stdio.h>
 #include <string.h>
-#include "../mlkem/src/fips202/fips202.h"
+#include "../../mlkem/src/fips202/fips202.h"
 
-#include "../mlkem/mlkem_native.h"
+#include "../../mlkem/mlkem_native.h"
 
 #if defined(_WIN64) || defined(_WIN32)
 #include <fcntl.h>
